Seems to work.

+++ src/helpers.cpp
@@ -4,7 +4,7 @@
 
 btScalar WrapAngle(const btScalar value)
 {
-    int i = value / 360.0;
-    i = (value < 0.0) ? (i - 1) : (i);
-    return value - 360.0 * i;
+    int i = static_cast<int>(value / 360.0);
+    i = (value < 0.0)?(i-1):(i);
+    return value - 360.0f * i;
 }
